audit/seccomp: Move parser to common folder
By moving the parser code to the common folder, we are allowing 3rd
party apps to use it without importing the code of kubectl-gadget.
